Extract logscan CLI/resume utilities into modules/logscan_resume.py
Second sub-phase of the logscan/trends breakup. Moves the CLI-string
and resume/recovery-suggestion formatting layer: flag manipulation,
config-path injection, recovery-command building, timestamp/duration
formatting, and the incomplete-run resume-explanation builder.

This layer sits between the cache primitives (sub-phase 1) and the
progress/content-analysis layer (sub-phases 3-4, still in
quickstart.py) - pure string/formatting logic with no monkeypatch
coupling between functions moving together in this phase, so no
deferred-import bridges were needed here (only _build_resume_library_scope
needed a noqa re-export, since two tests call it directly on qs_module
but nothing else in this phase calls it internally).

The functions live in quickstart.py across three separate regions with
gaps for the not-yet-moved imagemaid-analysis and progress-analysis
layers in between; extraction pulled functions from each region rather
than a single contiguous block.
